- contest tasks list

+++ Judge/Judge.Model/Contests/ContestTask.cs
@@ -4,7 +4,7 @@ namespace Judge.Model.Contests
 {
     public sealed class ContestTask
     {
-        public long ContestId { get; set; }
+        public int ContestId { get; set; }
         public string TaskName { get; set; }
         public Task Task { get; set; }
     }
